requirements.txt
Django==1.8
MySQL-python==1.2.5
Pillow==3.0.0
boto==2.38.0
django-social-auth==0.7.28
django-storages==1.1.8
django-tagulous==0.11.1
httplib2==0.9.2
oauth2==1.9.0.post1
python-openid==2.2.5
